+++ CHANGELOG.md
@@ -8,6 +8,7 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 ## [25.x.x]
 ### Changed
 - add explanations for the individual values in the feed information table
+- show error message from `opml` import in web-ui
 
 ### Fixed
 - fix proxy port removed if standard port for the protocol (#3027)